disallow negative locktimes in OutputOwners

+++ vms/avm/service_test.go
@@ -229,7 +229,7 @@ func TestServiceGetBalanceStrict(t *testing.T) {
 		Out: &secp256k1fx.TransferOutput{
 			Amt: 1337,
 			OutputOwners: secp256k1fx.OutputOwners{
-				Locktime:  now.Add(10 * time.Hour).Unix(),
+				Locktime:  uint64(now.Add(10 * time.Hour).Unix()),
 				Threshold: 1,
 				Addrs:     []ids.ShortID{addr},
 			},
@@ -421,7 +421,7 @@ func TestServiceGetAllBalances(t *testing.T) {
 		Out: &secp256k1fx.TransferOutput{
 			Amt: 1337,
 			OutputOwners: secp256k1fx.OutputOwners{
-				Locktime:  now.Add(10 * time.Hour).Unix(),
+				Locktime:  uint64(now.Add(10 * time.Hour).Unix()),
 				Threshold: 1,
 				Addrs:     []ids.ShortID{addr},
 			},
