+++ vms/secp256k1fx/keychain.go
@@ -121,9 +121,10 @@ func (kc *Keychain) Spend(out verify.Verifiable, time int64) (verify.Verifiable,
 	return nil, nil, fmt.Errorf("can't spend UTXO because it is unexpected type %T", out)
 }
 
-// Match attempts to match a list of addresses up to the provided threshold
+// Match attempts to match a list of addresses up to the provided threshold.
+// Assumes [time] is a Unix time > 0.
 func (kc *Keychain) Match(owners *OutputOwners, time int64) ([]uint32, []*secp256k1.PrivateKey, bool) {
-	if time < owners.Locktime {
+	if uint64(time) < owners.Locktime {
 		return nil, nil, false
 	}
 	sigs := make([]uint32, 0, owners.Threshold)
